uncompressed: pulled TileComponentInterleaveDecoder into its own file
